+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [Unreleased]
+
+### FIXED
+
+- [TD-6659] Fix non admin request grant for default permission
+- [TD-5149] Enriches implementation with deleted structures 
+
+### Changed
+
+- [TD-6602] Update td-cache, td-core and td-df-lib
+- [TD-6317] Improve performance of queries to retrieve nodes in linage
+
 ## [6.8.2] 2024-07-18
 
 ### Added